+++ src/components/BMDashboard/LogMaterials/LogMaterialsTable.jsx
@@ -1,4 +1,4 @@
-import { postMaterialLog, resetLogMaterialsResult, resetMaterialsByProjAndCheckInOut } from 'actions/inventoryMaterial';
+import { postMaterialLog, resetLogMaterialsResult, resetMaterialsByProjAndCheckInOut } from 'actions/bmdashboard/inventoryMaterial';
 import React, { useState } from 'react'
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
